+++ src/components/Header/components/HeaderSearchButton.tsx
@@ -1,5 +1,10 @@
 import FocusLock from "@chakra-ui/focus-lock";
-import { Box, chakra, useStyleConfig } from "@chakra-ui/react";
+import {
+  Box,
+  chakra,
+  useColorModeValue,
+  useStyleConfig,
+} from "@chakra-ui/react";
 import React, { useState, useRef } from "react";
 
 import Button from "../../Button/Button";
@@ -19,7 +24,8 @@ export interface HeaderSearchButtonProps {
 const HeaderSearchButton = chakra(
   ({ isMobile = false }: HeaderSearchButtonProps) => {
     const [isOpen, setIsOpen] = useState<boolean>(false);
-    const styles = useStyleConfig("HeaderSearchButton", { isOpen });
+    const isDarkMode = useColorModeValue(false, true);
+    const styles = useStyleConfig("HeaderSearchButton", { isOpen, isDarkMode });
     const buttonText = isMobile ? null : isOpen ? "Close" : "Search";
     const labelText = isOpen ? "Close Search" : "Open Search";
     const ref = useRef<HTMLDivElement>(null);