+++ src/theme/components/header/headerMobileNavButton.ts
@@ -17,6 +17,12 @@ const HeaderMobileNavButton = {
       },
     },
     _focus: headerFocus,
+    _dark: {
+      backgroundColor: isOpen ? "#2b2b2b" : "transparent",
+      svg: {
+        fill: "dark.ui.typography.heading",
+      },
+    },
   }),
 };